新闻推荐、商品推荐、新品推荐、Banner

+++ database/migrations/2018_06_14_155821_create_coupons_table.php
@@ -18,8 +18,8 @@ class CreateCouponsTable extends Migration
             $table->string('coupon_no');
             $table->string('name');
             $table->tinyInteger('type')->default(1)->comment('1满减 2抵扣');
-            $table->decimal('quota');
-            $table->decimal('satisfy')->nullable();
+            $table->integer('quota');
+            $table->integer('satisfy')->nullable();
             $table->integer('number');
             $table->integer('received')->default(0);
             $table->tinyInteger('is_meanwhile')->default(0);
